refactor(test): audit empty_and_bounded.rs — rename to empty.rs, rehome 3 tests
- Rename empty_and_bounded.rs → empty.rs (keep only empty-start-trie tests)
- Gate mod empty behind #[cfg(feature = "ethhash")] and remove per-test
  cfg attributes; empty roots only have a hash in ethhash mode
- Drop test_empty_start_trie_large_end_trie_multi_round (stress test,
  no unique code paths vs other tests)
- Drop test_empty_start_trie_single_edge (duplicative of bounded tests
  + bounds.rs)
- Add empty key (b"") to test_empty_start_trie_prefix_keys for root
  node value coverage
- Move test_bounded_range_with_existing_keys → bounds.rs as additional
  test_case (both edges with gap keys); remove duplicate exact-key case
- Move test_generator_uses_end_key_for_complete_proof → bounds.rs
- Move test_change_proof_prefix_key_deleted_in_end_root → edge_cases.rs
- Rename db_b → target; fix stale comment; remove redundant inline comments
